Rename fsstats to fsstat metricset
* Following rule that all metricsets are singular
* Ignore all data directories in beats
