+++ metricbeat/docs/modules/system/fsstat.asciidoc
@@ -2,8 +2,8 @@
 This file is generated! See scripts/docs_collector.py
 ////
 
-[[metricbeat-metricset-system-fsstats]]
-=== System FsStats MetricSet
+[[metricbeat-metricset-system-fsstat]]
+=== System Fsstat MetricSet
 
 The System FsStats MetricSet provides overall Filesystem statistics.
 
@@ -23,7 +23,7 @@ Below is an example document generated by this metricset.
         "hostname":"ruflin",
         "name":"ruflin"
     },
-    "metricset":"fsstats",
+    "metricset":"fsstat",
     "module":"system",
     "rtt":89,
     "system-fsstats":{
