+++ metricbeat/etc/fields.yml
@@ -898,10 +898,10 @@
                 The percentage of used disk space.
 
 
-        - name: fsstats
+        - name: fsstat
           type: group
           description: >
-            `system-fsstats` contains filesystem metrics aggregated from all mounted
+            `system.fsstat` contains filesystem metrics aggregated from all mounted
             filesystems.
           fields:
             - name: count
