+++ metricbeat/include/list.go
@@ -21,7 +21,7 @@ import (
 	_ "github.com/elastic/beats/metricbeat/module/system/cpu"
 	_ "github.com/elastic/beats/metricbeat/module/system/disk"
 	_ "github.com/elastic/beats/metricbeat/module/system/filesystem"
-	_ "github.com/elastic/beats/metricbeat/module/system/fsstats"
+	_ "github.com/elastic/beats/metricbeat/module/system/fsstat"
 	_ "github.com/elastic/beats/metricbeat/module/system/memory"
 	_ "github.com/elastic/beats/metricbeat/module/system/process"
 	_ "github.com/elastic/beats/metricbeat/module/zookeeper"
