+++ metricbeat/metricbeat.template-es2x.json
@@ -741,7 +741,7 @@
                 }
               }
             },
-            "fsstats": {
+            "fsstat": {
               "properties": {
                 "count": {
                   "type": "long"
