+++ metricbeat/metricbeat.template.json
@@ -711,7 +711,7 @@
                 }
               }
             },
-            "fsstats": {
+            "fsstat": {
               "properties": {
                 "count": {
                   "type": "long"
