+++ metricbeat/module/system/_beat/config.yml
@@ -1,5 +1,5 @@
 - module: system
-  metricsets: ["cpu", "core", "filesystem", "fsstats", "memory", "process"]
+  metricsets: ["cpu", "core", "filesystem", "fsstat", "memory", "process"]
   enabled: true
   period: 2s
   processes: ['.*']
