+++ metricbeat/module/system/fsstat/_beat/data.json
@@ -4,7 +4,7 @@
         "hostname":"ruflin",
         "name":"ruflin"
     },
-    "metricset":"fsstats",
+    "metricset":"fsstat",
     "module":"system",
     "rtt":89,
     "system-fsstats":{
