+++ metricbeat/module/system/fsstat/_beat/docs.asciidoc
@@ -1,3 +1,3 @@
-=== System FsStats MetricSet
+=== System Fsstat MetricSet
 
 The System FsStats MetricSet provides overall Filesystem statistics.
